Re-enable validation. Remove interpolation code.
Move code to plot missing data to a new notebook.
